Add check for MIME encoded messages
As far as I understand MIME-encoded message can be broken down into parts based upon its contents. While most mail programs add the plain text part as the first part this is not true for all programs. In fact mail2deck couldn't import the body when processing certain plain text mails with attachments - because the plain text part wasn't  at position 1.1.

With this fix imap_fetchbody requests sub-part 1 of part 1 and, if it is empty/ non-existent, it requests part 1.